Fix all examples
