+++ examples/auto-layout/src/ui.tsx
@@ -1,7 +1,7 @@
-import { uiWorker } from 'react-figma';
+import * as React from 'react';
+import { App } from './App';
 
-const handler = uiWorker({ fetch });
+import 'react-figma/rpc';
+import { render } from 'react-figma';
 
-onmessage = event => {
-    handler(event);
-};
+render(<App />);
