+++ examples/basic-hmr/src/code.tsx
@@ -1,3 +1,5 @@
-import '../../../src/rpc';
+import { setupMainThread } from 'react-figma/rpc';
 
 figma.showUI(__html__, { visible: false });
+
+setupMainThread();
