+++ examples/basic-hmr/src/ui.tsx
@@ -1,7 +1,8 @@
 import * as React from 'react';
-import '../../../src/rpc';
-import { render } from '../../../src';
 import { App } from './App';
+
+import 'react-figma/rpc';
+import { render } from 'react-figma';
 
 render(<App />);
