+++ examples/basic-hmr/webpack.config.js
@@ -24,7 +24,13 @@ module.exports = (env, argv) => ({
     },
 
     // Webpack tries these extensions for you if you omit the extension like "import './file'"
-    resolve: { extensions: ['.tsx', '.ts', '.jsx', '.js'] },
+    resolve: {
+        extensions: ['.tsx', '.ts', '.jsx', '.js'],
+        alias: {
+            'react-figma$': '../../../src',
+            'react-figma/rpc$': '../../../src/rpc'
+        }
+    },
 
     output: {
         filename: '[name].js',
