+++ examples/basic-hmr/webpack.ui.config.js
@@ -20,7 +20,13 @@ module.exports = (env, argv) => ({
     },
 
     // Webpack tries these extensions for you if you omit the extension like "import './file'"
-    resolve: { extensions: ['.tsx', '.ts', '.jsx', '.js'] },
+    resolve: {
+        extensions: ['.tsx', '.ts', '.jsx', '.js'],
+        alias: {
+            'react-figma$': '../../../src',
+            'react-figma/rpc$': '../../../src/rpc'
+        }
+    },
 
     output: {
         filename: '[name].js',
