+++ examples/basic/webpack.config.js
@@ -4,7 +4,8 @@ module.exports = configure({
     resolve: {
         extensions: ['.tsx', '.ts', '.jsx', '.js'],
         alias: {
-            'react-figma$': '../../../src' // oly for example, remove this at real app
+            'react-figma$': '../../../src',
+            'react-figma/rpc$': '../../../src/rpc'
         }
     }
 });
