+++ examples/design-system/src/code.tsx
@@ -1,4 +1,5 @@
-import * as React from 'react';
-import '../../../src/rpc';
+import { setupMainThread } from 'react-figma/rpc';
 
 figma.showUI(__html__, { visible: false });
+
+setupMainThread();
