+++ examples/design-system/src/ui.tsx
@@ -1,6 +1,7 @@
-import { App } from './App';
 import * as React from 'react';
-import '../../../src/rpc';
+import { App } from './App';
+
+import 'react-figma/rpc';
 import { render } from 'react-figma';
 
 render(<App />);
