--- examples/groups/manifest.json
@@ -1,7 +1,0 @@
-{
-  "name": "Groups",
-  "id": "741371398137186725",
-  "api": "1.0.0",
-  "main": "dist/code.js",
-  "ui": "dist/ui.html"
-}
