--- examples/groups/package.json
@@ -1,13 +1,0 @@
-{
-  "name": "basic",
-  "version": "0.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "webpack:watch": "webpack --watch"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-  }
-}
