--- examples/groups/src/ui.tsx
@@ -1,8 +1,0 @@
-import * as yoga from 'yoga-layout-prebuilt';
-import { uiWorker } from 'react-figma';
-
-const handler = uiWorker({ yoga, fetch });
-
-onmessage = event => {
-    handler(event);
-};
