--- examples/groups/tsconfig.json
@@ -1,21 +1,0 @@
-{
-  "compilerOptions": {
-    "outDir": "dist",
-    "module": "commonjs",
-    "target": "es5",
-    "lib": ["es5", "es6", "es7", "es2017", "dom"],
-    "sourceMap": true,
-    "allowJs": false,
-    "jsx": "react",
-    "moduleResolution": "node",
-    "rootDirs": ["src"],
-    "baseUrl": "src",
-    "skipLibCheck": true,
-    "declaration": true,
-    "paths": {
-      "react-figma": ["../../../src/index.ts"]
-    }
-  },
-  "include": ["src/**/*", "../../figma.d.ts", "../../jsx-interfaces.d.ts", "../../images.d.ts"],
-  "exclude": ["node_modules", "dist", "**/__tests__/**"]
-}
