--- examples/groups/webpack.config.js
@@ -1,10 +1,0 @@
-var configure = require('react-figma-webpack-config');
-
-module.exports = configure({
-    resolve: {
-        extensions: ['.tsx', '.ts', '.jsx', '.js'],
-        alias: {
-            'react-figma$': '../../../src' // oly for example, remove this at real app
-        }
-    }
-});
