+++ examples/styled-components/package.json
@@ -25,19 +25,9 @@
   "license": "ISC",
   "dependencies": {
     "chroma-js": "^2.1.0",
-    "react": "^16.9.0",
-    "react-figma": "0.0.57",
+    "react-dom": "^16.13.1",
+    "react-native-web": "^0.12.2",
     "react-primitives": "^0.8.1",
-    "styled-components": "^5.0.1",
-    "yoga-layout-prebuilt": "^1.9.3"
-  },
-  "devDependencies": {
-    "html-webpack-inline-source-plugin": "^0.0.10",
-    "html-webpack-plugin": "^3.2.0",
-    "react-figma-webpack-config": "0.0.5",
-    "ts-loader": "^6.0.4",
-    "typescript": "^3.5.3",
-    "webpack": "^4.39.2",
-    "webpack-cli": "^3.3.7"
+    "styled-components": "^5.0.1"
   }
 }
