+++ examples/styled-components/src/styled.ts
@@ -1,18 +1,6 @@
-// @ts-ignore FIXME:
-import { Platform } from 'react-primitives';
-
-let styled, ThemeProvider;
-
-import styledW, { ThemeProviderW } from 'styled-components';
 import styledP, { ThemeProviderP } from 'styled-components/primitives';
-
-if (Platform.OS === 'figma') {
-    styled = styledP;
-    ThemeProvider = ThemeProviderP;
-} else {
-    styled = styledW;
-    ThemeProvider = ThemeProviderW;
-}
+const styled = styledP;
+const ThemeProvider = ThemeProviderP;
 
 export { ThemeProvider };
