+++ examples/styled-components/webpack.config.js
@@ -1,11 +1,11 @@
-const path = require('path');
-const configure = require('react-figma-webpack-config');
+var configure = require('react-figma-webpack-config');
 
 module.exports = configure({
     resolve: {
+        extensions: ['.tsx', '.ts', '.jsx', '.js'],
         alias: {
-            react: path.resolve(__dirname, './node_modules/react/'),
-            'react-figma': path.resolve(__dirname, './node_modules/react-figma/')
+            'react-figma$': '../../../src',
+            'react-figma/rpc$': '../../../src/rpc'
         }
     }
 });
